GH_ISSUE_906: split admin/handler.go by domain
The single 1219-line handler.go mixed route registration, auth
middleware, status/observability endpoints, cache control, replication,
backend lifecycle, encryption key rotation + bulk rewrite, integrity
ops, and helper functions. Review and navigation were getting harder
per #906; the admin surface is expected to keep growing.

Splits into seven domain-focused siblings; handler.go shrinks to the
Handler/Deps types + constructor + internalError helper:

  - handler_routes.go: Register + requireToken middleware.
  - handler_status.go: status, reload-status, object-locations,
    cleanup-queue, usage-flush, log-level, workers + WorkerHealth.
  - handler_cache.go: cache flush/stats/invalidate-key/-prefix +
    cacheDisabledReason/writeCacheDisabled.
  - handler_replication.go: ReplicateResult + Replicate +
    handleReplicate + over-replication status/clean.
  - handler_backends.go: drain start/progress/cancel +
    handleRemoveBackend with HMAC-signed two-phase purge confirm.
  - handler_encryption.go: key rotation + bulk encrypt/decrypt
    (bulkRewriteRow/Op + runBulkRewriteCounts + processBulkLocation
    generics moved together).
  - handler_integrity.go: Scrub + BackfillChecksums (typed + handler
    forms) + handleReconcile.

Pure structural move - no behaviour change, no test changes.
