+++ .version
@@ -1,1 +1,1 @@
-v0.60.5
+v0.60.6
